# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CircleTel API
  description: |
    CircleTel B2B/B2C ISP Platform API

    This API provides access to CircleTel's fibre connectivity services including:
    - Coverage checking and package selection
    - Order management and payment processing
    - Customer dashboard and service management
    - Admin operations and reporting

    ## Authentication

    The API uses different authentication methods depending on the context:

    ### Consumer/Partner Authentication
    - Uses Supabase Auth with JWT tokens
    - Token passed via `Authorization: Bearer <token>` header
    - Session cookies as fallback for SSR

    ### Admin Authentication
    - Requires admin user with appropriate RBAC permissions
    - Uses service role for cross-table operations

    ### Webhook Authentication
    - HMAC-SHA256 signature verification
    - IP whitelist validation

    ## Rate Limiting

    Standard rate limits apply:
    - 100 requests/minute for authenticated users
    - 20 requests/minute for unauthenticated endpoints

    ## Error Responses

    All errors follow the standard format:
    ```json
    {
      "success": false,
      "error": "ErrorType",
      "message": "Human-readable message",
      "code": "ERR_CODE",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: CircleTel Development Team
    email: dev@circletel.co.za
    url: https://www.circletel.co.za
  license:
    name: Proprietary
...
servers:
...
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token

    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie

    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: HMAC-SHA256 signature of request body
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
        message:
          type: string
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - code
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Error type name
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Error code (e.g., AUTH_001, VAL_002)
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        perPage:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # Coverage Types
...
    ServicePackage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "SkyFibre 100Mbps"
        speed_down:
          type: integer
          example: 100
        speed_up:
          type: integer
          example: 50
        price:
          type: number
          format: decimal
          example: 799.00
        installation_fee:
          type: number
          format: decimal
          example: 1500.00
        service_type:
          type: string
          enum: [fibre_consumer, fibre_business, wireless, lte]
        provider:
          type: string
        features:
          type: array
          items:
            type: string
        active:
          type: boolean

    # Order Types
...
paths:
...
  /coverage/packages:
    get:
      tags:
        - Coverage
      summary: Get available packages for a lead
      description: Retrieves available service packages based on coverage lead information
      operationId: getPackages
      parameters:
        - name: leadId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Coverage lead ID from previous coverage check
      responses:
        '200':
          description: Packages retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          packages:
                            type: array
                            items:
                              $ref: '#/components/schemas/ServicePackage'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Order Endpoints
...